Saniya-Quillzy-Editor/smart-editor/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pymongo==4.8.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
requests==2.32.3
